pyproject.toml
[project]
name = "crewai-linkedin-post"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.12.3",
    "crewai>=0.119.0",
    "crewai-tools>=0.44.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain-core>=0.3.59",
    "langchain-mistralai>=0.2.10",
    "langchain-openai>=0.2.14",
    "lxml==5.2.1",
    "pydantic>=2.11.4",
    "python-dotenv==1.0.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "selenium==4.19.0",
    "setuptools>=80.4.0",
    "streamlit>=1.45.0",
    "tiktoken>=0.9.0",
    "undetected-chromedriver>=3.5.5",
    "uvicorn>=0.34.2",
]
